pyproject.toml
[project]
name = "followthefootprints"
version = "0.1.0"
description = "Stock demand zone scanner using the Follow The Footprints strategy"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "yfinance>=0.2.50",
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "lxml>=6.0.2",
    "requests>=2.32.5",
    "python-dotenv>=1.0.0",
]

[project.scripts]
ftf = "followthefootprints.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/followthefootprints"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
